C11-159: extract debug.* socket handlers (gating preserved)
DebugHandlers.swift: debug.terminals stays ungated (available in release);
all 34 other debug.* cases and their v2Debug* methods stay under #if DEBUG,
exactly as before. Removed the now-empty #if DEBUG blocks left in the switch
and method regions. TerminalController.swift: 14145 -> 13222 LOC.
